mypy part 1

+++ src/entity/logics/brain.py
@@ -6,7 +6,7 @@
 #  By: anacharp, roandrie                        +#+  +:+       +#+         #
 #                                              +#+#+#+#+#+   +#+            #
 #  Created: 2026/05/29 14:10:28 by roandrie        #+#    #+#               #
-#  Updated: 2026/06/15 08:48:36 by roandrie        ###   ########.fr        #
+#  Updated: 2026/06/15 09:15:55 by roandrie        ###   ########.fr        #
 #                                                                           #
 # ************************************************************************* #
 
@@ -64,7 +64,7 @@ class EnemyBrain():
             self.enemy.calculator.get_pixel_to_grid_entity(self.enemy)
         )
 
-        open_walls: dict[tuple[float, float], tuple[int, int]] = (
+        open_walls: dict[tuple[int, int], tuple[int, int]] = (
             self.enemy.calculator.check_open_wall(
                 int(conv_x), int(conv_y), self.enemy.maze_bitmap))
